fix bug on Key

+++ source/app/controller/key.go
@@ -1,48 +1,64 @@
 package controller
 
 import (
+	"fmt"
 	"github.com/gin-gonic/gin"
-	"net/http"
+	"yscloudeBack/source/app/db"
 	"yscloudeBack/source/app/model"
-	"yscloudeBack/source/app/utils"
 )
 
-func RegisterKey(db *model.DbManager) gin.HandlerFunc {
+// 注册key格式:
+//
+//	{
+//		"usage":"导入",
+//		"num":1,
+//		"fileGroupName":""
+//	}
+func RegisterKey(db *db.DbManager) gin.HandlerFunc {
 	return func(ctx *gin.Context) {
-		key, err := utils.GenerateRandomKey()
+		var form struct {
+			Usage         string `json:"usage"`
+			Num           int    `json:"num"`
+			FileGroupName string `json:"file_group_name"`
+		}
+		code, err2 := model.BindStruct(ctx, &form)
+		if err2 != nil {
+			model.BackError(ctx, code)
+			return
+		}
+
+		key, err := model.NewKey(form.Usage, 1, form.FileGroupName)
 		if err != nil {
-			model.BackError(ctx, model.CodeGetKeyFalse)
+			model.BackError(ctx, model.CodeCreateKeyFalse)
 			return
 		}
 		err = db.AddKey(key)
 		if err != nil {
+			fmt.Println(err)
 			model.BackError(ctx, model.CodeInvalidKey)
 			return
 		}
-		ctx.JSON(http.StatusOK, gin.H{
-			"code": model.CodeSuccess,
-			"msg":  model.CodeSuccess.Msg(),
-		})
+		model.BackSuccess(ctx, key)
 	}
 }
 
 // 获取keys
-func GetKey(db *model.DbManager) gin.HandlerFunc {
+func GetKey(db *db.DbManager) gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		keys, err := db.GetAllKeys()
 		if err != nil {
 			model.BackError(ctx, model.CodeUnknowError)
 			return
 		}
-		key_values := []string{}
-		for _, v := range keys {
-			key_values = append(key_values, v.Value)
-		}
-		model.BackSuccess(ctx, key_values)
+		//key_values := []string{}
+		//for _, v := range keys {
+		//	key_values = append(key_values, v.Value)
+		//}
+		model.BackSuccess(ctx, keys)
 		return
 	}
 }
-func DelKey(db *model.DbManager) gin.HandlerFunc {
+func DelKey(db *db.DbManager) gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		var form struct {
 			DelKey string `form:"del_key" binding:"required" json:"del_key"`
@@ -53,15 +69,22 @@ func DelKey(db *model.DbManager) gin.HandlerFunc {
 			model.BackError(ctx, model.CodeInvalidKey)
 			return
 		}
+		isFind, key := db.CheckKey(form.DelKey)
+		if key != nil {
+			model.BackError(ctx, model.CodeUnknowError)
+			return
+		}
+		if !isFind {
+			model.BackError(ctx, model.CodeGetKeyFalse)
+			return
+		}
+
 		err := db.DeleteKey(form.DelKey)
 		if err != nil {
 			model.BackError(ctx, model.CodeUnknowError)
 			return
 		}
-		ctx.JSON(http.StatusOK, gin.H{
-			"code": model.CodeSuccess,
-			"msg":  model.CodeSuccess.Msg(),
-		})
+		model.BackSuccess(ctx, nil)
 		return
 	}
